Separates HomeViewModel from AppViewModel, improves a few layouts.
