+++ app/src/main/java/com/myportfolio/mymovieapp/ui/adapters/MovieListAdapter.kt
@@ -5,6 +5,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.cardview.widget.CardView
 import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
@@ -13,9 +14,13 @@ import com.myportfolio.mymovieapp.model.Media
 import com.squareup.picasso.Picasso
 
 
-class MovieListAdapter:
-    RecyclerView.Adapter<MovieListAdapter.MovieListViewHolder>() {
+class MovieListAdapter: RecyclerView.Adapter<MovieListAdapter.MovieListViewHolder>() {
 
+    enum class MediaPosterSize {
+        SMALL, BIG
+    }
+
+    private var mediaPosterSize: MediaPosterSize = MediaPosterSize.SMALL
     private var onItemClicked: ((Media) -> Unit)? = null
 
     private val differCallBack = object: DiffUtil.ItemCallback<Media>() {
@@ -26,9 +31,7 @@ class MovieListAdapter:
         override fun areContentsTheSame(oldItem: Media, newItem: Media): Boolean {
             return oldItem == newItem
         }
-
     }
-
     private val differ = AsyncListDiffer(this, differCallBack)
 
     inner class MovieListViewHolder(view: View): RecyclerView.ViewHolder(view) {
@@ -41,19 +44,31 @@ class MovieListAdapter:
         val layout = LayoutInflater
             .from(parent.context)
             .inflate(R.layout.movie_list_item, parent, false)
+        if (mediaPosterSize == MediaPosterSize.BIG) {
+            val mediaCard = layout.findViewById<CardView>(R.id.movie_image_card)
+            mediaCard.layoutParams.width = 492
+            mediaCard.layoutParams.height = 640
+            mediaCard.radius = 20f
+        }
         // Setup custom accessibility delegate to set the text read
         return MovieListViewHolder(layout)
     }
 
     override fun onBindViewHolder(holder: MovieListViewHolder, position: Int) {
         val media: Media = differ.currentList[position]
-        if (media.smallPoster.trim().isEmpty()) {
+        val mediaPoster: String = if (mediaPosterSize == MediaPosterSize.BIG) {
+            media.mediumPoster
+        } else {
+            media.smallPoster
+        }
+        if (mediaPoster.trim().isEmpty()) {
             holder.image.setImageResource(R.drawable.ic_launcher_background)
         } else {
             Picasso.get()
-                .load(media.smallPoster)
+                .load(mediaPoster)
                 .into(holder.image)
         }
+
         holder.movieTitle.text = media.mediaName
         holder.movieReleaseYear.text = media.releaseYear.toString()
         holder.itemView.setOnClickListener {
@@ -71,6 +86,10 @@ class MovieListAdapter:
 
     fun setOnMediaClickListener(listener: (Media) -> Unit) {
         onItemClicked = listener
+    }
+
+    fun setMediaPosterSizeToBig() {
+        mediaPosterSize = MediaPosterSize.BIG
     }
 
 }
