+++ app/src/main/java/com/myportfolio/mymovieapp/util/Constants.kt
@@ -1,0 +1,8 @@
+package com.myportfolio.mymovieapp.util
+
+object Constants {
+
+    const val movieTypeId = "movie"
+    const val seriesTypeId = "tv"
+
+}
