+++ app/src/main/res/layout/fragment_category.xml
@@ -13,10 +13,12 @@
         android:id="@+id/category_fragment_title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:padding="16dp"
+        android:paddingHorizontal="8dp"
+        android:paddingVertical="20dp"
         android:text="Category Title"
         android:textSize="28sp"
         android:textStyle="bold"
+        android:textColor="@color/dark_grey"
         />
 
     <androidx.recyclerview.widget.RecyclerView
@@ -27,7 +29,5 @@
         app:spanCount="2"
         tools:listitem="@layout/movie_list_item"
         />
-
-
 
 </androidx.appcompat.widget.LinearLayoutCompat>
