+++ app/src/main/res/layout/movie_list.xml
@@ -13,7 +13,8 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="Category"
-            android:padding="8dp"
+            android:paddingHorizontal="16dp"
+            android:paddingVertical="4dp"
             style="@style/Widget.MyMovieApp.CategoryTitleTextView"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
@@ -23,7 +24,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="horizontal"
-            android:padding="8dp"
+            android:paddingTop="8dp"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
             app:layout_constraintTop_toBottomOf="@id/category_title"
             tools:listitem="@layout/movie_list_item">
