+++ app/src/main/res/layout/movie_list_item.xml
@@ -8,14 +8,23 @@
     android:layout_margin="8dp"
     >
 
-    <ImageView
-        android:id="@+id/movie_image"
-        android:layout_height="100dp"
-        android:layout_width="80dp"
-        android:contentDescription="@string/movie_image"
-        android:scaleType="centerCrop"
-        tools:src="@drawable/ic_launcher_background"
-        />
+    <androidx.cardview.widget.CardView
+        android:id="@+id/movie_image_card"
+        android:layout_width="92dp"
+        android:layout_height="120dp"
+        app:cardCornerRadius="4dp"
+    >
+
+        <ImageView
+            android:id="@+id/movie_image"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:contentDescription="@string/movie_image"
+            android:scaleType="centerCrop"
+            tools:src="@drawable/ic_launcher_background"
+            />
+
+    </androidx.cardview.widget.CardView>
 
     <TextView
         android:id="@+id/movie_title"
@@ -24,6 +33,7 @@
         android:text="Titulo"
         android:textStyle="bold"
         android:textSize="16sp"
+        android:textColor="@color/dark_grey"
         android:lines="1"
         android:ellipsize="end"
         />
