+++ app/src/main/res/values/colors.xml
@@ -2,6 +2,7 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="dark_grey">#FF3D3D3D</color>
     <color name="category_title_fill">#BC8034</color>
     <color name="movie_genre_text_color">#545454</color>
 </resources>
