+++ app/src/main/res/values/styles.xml
@@ -8,5 +8,6 @@
     <style name="Widget.MyMovieApp.HomeTitleTextView" parent="Widget.AppCompat.TextView">
         <item name="android:textSize">24sp</item>
         <item name="android:textStyle">bold</item>
+        <item name="android:textColor">@color/dark_grey</item>
     </style>
 </resources>
